fix: more narrowing conversion errors for macOS (LBIE_Mesher, Table)

+++ src/ColorTable2/Table.cpp
@@ -721,9 +721,9 @@ namespace CVCColorTable
   {
      if( !_dirtyContourTree ) return;
   
-     int dim[3] = { _contourVolume.XDim(),
-                    _contourVolume.YDim(),
-                    _contourVolume.ZDim() };
+     int dim[3] = { static_cast<int>(_contourVolume.XDim()),
+                    static_cast<int>(_contourVolume.YDim()),
+                    static_cast<int>(_contourVolume.ZDim()) };
 
      _contourTreeVertices.clear();
      _contourTreeEdges.clear();
